+++ examples/web/schema-and-deployments/create_table_single_node.ts
@@ -1,0 +1,13 @@
+import { createClient } from '@clickhouse/client-web'
+
+// A single ClickHouse node - for example, as in our `docker-compose.yml`
+const client = createClient()
+await client.command({
+  query: `
+    CREATE TABLE IF NOT EXISTS clickhouse_js_create_table_example
+    (id UInt64, name String)
+    ENGINE MergeTree()
+    ORDER BY (id)
+  `,
+})
+await client.close()